[FedCM] Do not crash for pending disconnect/getUserInfo
If the FederatedAuthRequestImpl is invoked and there are pending
requests, the destructor will clear out the FederatedAuthRequestImpl
members. This will in turn invoke the members' destructors, which then
complete their requests. However, to do so they run the completion
callback which is currently expecting the members to be present. This
CL fixes this by removing that expectation, and adding WPTs that would
have crashed in Chrome before this CL.

Bug: 1504569
Change-Id: I3327398740ee641d43a2a4b93420ac8ff56ae8fc