Grid items with align/justify-self should participate in baseline alginment
When a grid item is set to align itself to the baseline (first or last)
then it should participate in baseline alignment if the item's size is
not dependent on the size of the track and vice versa.

Also added a test to check for last baseline alignment in addition to
first baseline alignment on empty blocks.

Differential Revision: https://phabricator.services.mozilla.com/D157345

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1609403
gecko-commit: 3d0a5f6fd8b4a0537cb85ae502a4be6cafec1c23
gecko-reviewers: emilio, layout-reviewers