[A11y] Simplification -- don't create tree holes for layout updates
When the layout tree changes, invalidate the related a11y subtree
completely.

This removes one more call site for RepairChildrenOfIncludedParent(),
leaving only one caller remaining, which we hope to remove in a
follow-up.

Also fixes a bug, where the descendants of a textarea were missing
a parent.

Fixed: 1503069
Change-Id: If2176b5d013a0c2e0ecbd581a99ac45e4619ceda
Cq-Do-Not-Cancel-Tryjobs: true