[LineBreaker] Support rewinding floats in `nowrap`
This patch supports rewinding floats in `nowrap` or `pre`.

In Blink, when NG was first implemented, rewinding floats
was harder, as it had separated lists for floats. So rewinding
stopped at the last float. This wasn't really correct, but it
passed all existing tests.

This patch changes `Rewind()` to actually rewind positioned
floats when necessary.

The spec isn't very clear about this case. When Gecko fixed
their reported cases 5 years ago[1], they filed a spec
issue[2] and added several tests[3].

This patch changes the rendering of `float-nowrap-4.html`, one
of the tests by Mozilla, so this patch adds a new reference.
PS18 can fix this issue without changing the behavior for
`float-nowrap-4.html`, but logically speaking, this behavior
looks more correct.

The change is under a runtime flag `RewindFloats`, enabled by
default as a kill-switch in case this change turned out to be
not web-compatible.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=488725#c38
[2] https://github.com/w3c/csswg-drafts/issues/3038
[3] https://wpt.fyi/results/css/CSS2/floats?label=master&label=experimental&aligned&q=float-nowrap-

Bug: 1499290
Change-Id: I9f07fb487eaecf322c43df17e240afc5bc085560