Missing ::first-letter style update for size queries
Style for ::first-letter pseudo elements are updated as late as possible
to allow nodes to re-attach to make sure any first-letter text node
changes are included when computing style for the first rendered letter.

When the StyleRecalcChange were already marked for ReattachLayoutTree()
when reaching the ::first-letter update, we assumed some ancestor of the
::first-letter element would be re-attached, and the ::first-letter
style being updated in Element::AttachLayoutTree() for the originating
element. But, if the originating element was a size container, and we
skipped style recalc for its descendants, we would mark the stored
StyleRecalcChange for re-attachment of the children when resuming, even
if the originating element itself would be re-attached.

Instead, check that we also did not suppress style recalc for the
originating element for the ::first-letter before skipping its style
update during style recalc.

Bug: 1505250
Change-Id: I8d45c219e0d8c2a2d8ab6e0ec6149d9f34366fad